bug fix from 20.08
- fixed animation direction logic
- fixed rounding logic with min step
- fixed texts overlapping/truncating

+++ data/src/main/java/com/example/data/entity/QuoteEntity.kt
@@ -1,12 +1,14 @@
 package com.example.data.entity
 
+import java.math.BigDecimal
+
 
 data class QuoteEntity(
     val ticker: String? = null,
-    val priceChangeInPoints: Double? = null,
-    val latestTradePrice: Double? = null,
+    val priceChangeInPoints: BigDecimal? = null,
+    val latestTradePrice: BigDecimal? = null,
     val exchangeOfLatestTrade: String? = null,
-    val minStep: Double? = null,
+    val minStep: BigDecimal? = null,
     val name: String? = null,
-    val priceChangeByPercentage: Double? = null,
+    val priceChangeByPercentage: BigDecimal? = null,
 )
